Try to fix arm64 builds and build failures
Use native arm64 workers and also allow lib3mf to be built on arm64.

Currently all workers have 4 cores, but vcpkg uses (vcpu+1) cores for
builds. However, one of the build failures was due to the Action agent losing
connectivity to the GH servers. Hopefully this will help not
overexhausting the CPU of the worker.